Database tests fixed and bug in POST and PUT methods fixed

+++ src/main/kotlin/com/polarboookshop/catalogservice/domain/BookService.kt
@@ -4,20 +4,20 @@ import org.springframework.stereotype.Service
 import kotlin.jvm.Throws
 
 @Service
-class BookService {
-    lateinit var bookRepository: BookRepository
-
-    constructor(bookRepository: BookRepository) {
-        this.bookRepository = bookRepository
-    }
+class BookService(var bookRepository: BookRepository) {
 
     fun viewBookList(): Iterable<Book> {
         return bookRepository.findAll()
     }
 
     @Throws(BookNotFoundException::class)
-    fun viewBookDetails(isbn: String): Book? {
-        return bookRepository.findByIsbn(isbn)?.get()
+    fun viewBookDetails(isbn: String): Book {
+        if (bookRepository.existsByIsbn(isbn)) {
+            return bookRepository.findByIsbn(isbn)!!.get()
+        } else {
+            throw BookNotFoundException(isbn)
+        }
+
     }
 
     fun addBookToCatalog(book: Book): Book {
@@ -28,19 +28,25 @@ class BookService {
     }
 
     fun removeBookFromCatalog(isbn: String): Boolean {
-        return bookRepository.deleteByIsbn(isbn)
+        if (bookRepository.existsByIsbn(isbn)) {
+            return bookRepository.deleteByIsbn(isbn)
+        } else {
+            throw BookNotFoundException(isbn)
+        }
+
     }
 
     fun editBookDetails(isbn: String, book: Book): Book {
         return if (bookRepository.findByIsbn(isbn)!!.isPresent) {
             var bookToUpdate = bookRepository.findByIsbn(isbn)
-            bookToUpdate!!.get().id = book.id
+            bookToUpdate!!.get().id = bookToUpdate.get().id
+            bookToUpdate.get().isbn = bookToUpdate.get().isbn
             bookToUpdate.get().title = book.title
             bookToUpdate.get().author = book.author
             bookToUpdate.get().price = book.price
-            bookToUpdate.get().createdDate = book.createdDate
-            bookToUpdate.get().lastModifiedDate = book.lastModifiedDate
-            bookToUpdate.get().version = book.version
+            bookToUpdate.get().createdDate = bookToUpdate.get().createdDate
+            bookToUpdate.get().lastModifiedDate = bookToUpdate.get().lastModifiedDate
+            bookToUpdate.get().version = bookToUpdate.get().version
             bookRepository.save(bookToUpdate.get())
         } else {
             addBookToCatalog(book)
